+++ graph_attention_student/keras.py
@@ -12,12 +12,21 @@ an argument to a keras ``custom_object_scope`` context:
         model = ks.models.load_model('model/path')
 
 """
+import tensorflow.keras as ks
+
 from graph_attention_student.models.megan import Megan
+from graph_attention_student.models.megan import FidelityMegan
 from graph_attention_student.training import NoLoss
 from graph_attention_student.training import ExplanationLoss
 
 CUSTOM_OBJECTS = {
     'Megan': Megan,
+    'FidelityMegan': FidelityMegan,
     'NoLoss': NoLoss,
     'ExplanationLoss': ExplanationLoss
 }
+
+
+def load_model(path: str):
+    with ks.utils.custom_object_scope(CUSTOM_OBJECTS):
+        return ks.models.load_model(path)